requirements.txt
numpy~=1.21.5
opencv-python~=4.8.1.78
basicsr~=1.2.0+cc63400
torchvision~=0.12.0
pyyaml~=6.0.1
dlib~=19.24.2
scikit-image~=0.19.3
lmdb~=1.4.1
tqdm~=4.66.1
requests~=2.28.1
scipy~=1.7.3
h5py~=3.8.0
einops~=0.6.1
windnd~=1.0.7
future~=0.18.3
fvcore~=0.1.6
natsort~=8.4.0
pillow~=9.3.0
matplotlib~=3.5.3
ultralytics~=8.0.145